cleanup: refactor ssm code for packing naming without underscore
