+++ pkg/resource/ssm/node/printer.go
@@ -1,4 +1,4 @@
-package ssm_node
+package node
 
 import (
 	"io"
