+++ pkg/resource/ssm/session/printer.go
@@ -1,4 +1,4 @@
-package ssm_session
+package session
 
 import (
 	"io"
